etl/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
anyio==4.4.0
apache-airflow==2.9.2
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.14.0
apache-airflow-providers-fab==1.1.1
apache-airflow-providers-ftp==3.9.1
apache-airflow-providers-http==4.11.1
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.1
apispec==6.6.1
argcomplete==3.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
blinker==1.8.2
cachelib==0.9.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==42.0.8
Deprecated==1.2.14
dill==0.3.8
dnspython==2.6.1
docutils==0.21.2
email_validator==2.1.1
exceptiongroup==1.2.1
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.6.0
google-re2==1.1.20240601
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
limits==3.12.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.3.0
multidict==6.0.5
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
ordered-set==4.1.0
packaging==24.1
pathspec==0.12.1
pendulum==3.0.0
pluggy==1.5.0
prison==0.2.1
protobuf==4.25.3
psutil==5.9.8
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.7.3
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.1
rpds-py==0.18.1
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.1
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==2.2.1
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.19.2
